refactor: reorganize test
remove empty tests
separate groups to different files
remove initial widget tests
